chore: replace and add some necessary fonts
XeLaTeX uses Fandol fonts as fallbacks of CTeX Chinese fonts, which are
not very same as the required SimSong, SimHei, etc. So we manually
specify some fonts and make them consistent with the requirement of the
contest as much as possible. But please note that SImSong and other
Microsoft Chinese fonts don't have the corresponding bold and italic
variations. So we need `AutoFakeBold` in `cumcmthesis.cls` as well.